+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -321,7 +321,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSMDBSWP6195JZCBD2T9KA9E"
+        "01KSME7X4HBVX0QVTPG5TP1XNP"
     )
     r"""The unique identifier of the data source"""