+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSMDBSWKZXPWQN3EFBMP1KJX"
+        "01KSME7X4FGWN8909986M69SHA"
     )
     r"""The unique identifier of the data source"""