+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSMDBSWNAERCZ4A2QQK178DZ"
+        "01KSME7X4GNTWD6THSAAT2CEYD"
     )
     r"""The unique identifier of the data source"""