+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSMDBSWRNVNFWJR47VDPSVNW"
+        "01KSME7X4KHPDVNEJQ6PJ5ZJXV"
     )
     r"""The unique identifier of the data source"""