+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -292,7 +292,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KSMDBSRVJ9ZQ8VWFZ0PXRJV8"
+    id: Optional[str] = "01KSME7X18AEZZZYG8KZTQJ3P1"
 
     description: Optional[str] = None
 
@@ -1357,7 +1357,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KSMDBSRMWH7PFJF2XPG5QB1R"
+        "tool_01KSME7X115EBRS700RSJV3BRQ"
     )
 
     display_name: Optional[str] = None
@@ -1479,7 +1479,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KSMDBSRMAD3DCTH7A7QRG6RW"
+    id: Optional[str] = "01KSME7X109X3EDH1K44P9ZHQM"
 
     description: Optional[str] = None
 
@@ -1623,7 +1623,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KSMDBSRJ7QE95MB0KBDZD69R"
+        "tool_01KSME7X0Y38A3TW2828VPN9QT"
     )
 
     display_name: Optional[str] = None
@@ -1914,7 +1914,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KSMDBSRFVWA1933W7DDSBNS6"
+        "tool_01KSME7X0X3SFGC1925BQ8A9AH"
     )
 
     display_name: Optional[str] = None
@@ -2099,7 +2099,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KSMDBSRD5DXBMSWRH76KVB4T"
+        "tool_01KSME7X0VA3CYPQ9SW39Z7ZQF"
     )
 
     display_name: Optional[str] = None
@@ -2288,7 +2288,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KSMDBSRCA52R7932TPP0AKH6"
+        "tool_01KSME7X0TMAK00V0KAGAVWKSJ"
     )
 
     display_name: Optional[str] = None